Attempted build fix on win x64
Failing compilation on chromium.lkgr. Been failing for a while (~week).

Warning is:

c:\b\build\slave\win_x64\build\src\media\cast\rtcp\rtcp_sender_unittest.cc(492) :error C2220: warning treated as error - no 'object' file generated
c:\b\build\slave\win_x64\build\src\media\cast\rtcp\rtcp_sender_unittest.cc(492) : warning C4267: 'argument' : conversion from 'size_t' to 'uint16', possible loss of data

R=cpu@chromium.org, hubbe@chromium.org, jschuh@chromium.org, kbr@chromium.org

Review URL: https://codereview.chromium.org/100643004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@239589 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: d83bc30bee578764900e7297211da5d1edb7cbe1

+++ media/cast/rtcp/rtcp_sender_unittest.cc
@@ -489,7 +489,8 @@ TEST_F(RtcpSenderTest, RtcpReceiverReportWithOversizedFrameLog) {
       kRtcpMaxReceiverLogMessages, kTimeBaseMs);
 
   for (size_t i = 0; i < kRtcpMaxReceiverLogMessages; ++i) {
-    p.AddReceiverEventLog(kLostPacketId1, 6, kTimeDelayMs * i);
+    p.AddReceiverEventLog(
+        kLostPacketId1, 6, static_cast<uint16>(kTimeDelayMs * i));
   }
 
   test_transport_.SetExpectedRtcpPacket(p.Packet(), p.Length());
